# controlplane/api/openapi.yaml
openapi: 3.1.0
info:
  title: dlh-controlplane
  version: 0.1.0
  description: |
    Phase C API for dlh-test-fw. Includes submission endpoints (POST
    /api/runs, DELETE /api/runs/{id}) and internal chaos proxy endpoints.
servers:
  - url: /
paths:
  /healthz:
    get:
      operationId: getHealthz
      security: []
      responses:
        "200":
          description: OK
  /readyz:
    get:
      operationId: getReadyz
      security: []
      responses:
        "200":
          description: OK
        "503":
          description: Not ready
  /api/scenarios:
    get:
      operationId: listScenarios
      responses:
        "200":
          description: scenario catalog
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Scenario"
  /api/scenarios/{id}:
    get:
      operationId: getScenario
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: scenario detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scenario" }
        "404":
          description: not found
  /api/runs:
    get:
      operationId: listRuns
      parameters:
        - in: query
          name: scenario
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: since
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: run history
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
    post:
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "202":
          description: accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400":
          description: invalid request
        "404":
          description: scenario not found
  /api/runs/{id}:
    get:
      operationId: getRun
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: run detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunDetail" }
        "404":
          description: not found
    delete:
      operationId: cancelRun
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "202":
          description: cancellation accepted
        "404":
          description: not found
  /api/runs/{id}/events:
    get:
      operationId: streamRunEvents
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: server-sent events stream
          content:
            text/event-stream:
              schema:
                type: string
  /internal/chaos:
    post:
      operationId: createChaos
      description: |
        Called by Argo Workflow `http` template steps from the chaos WTs.
        Body is the Unstructured chaos CR (apiVersion + kind + metadata + spec).
        Auth is via the X-Internal-Token header carrying the controlplane's
        shared secret — NOT OIDC.
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChaosResource" }
      responses:
        "201":
          description: chaos resource created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChaosResourceRef" }
        "400":
          description: invalid body
        "401":
          description: bad internal token
        "500":
          description: cluster API error
  /internal/chaos/{ref}:
    delete:
      operationId: deleteChaos
      parameters:
        - in: path
          name: ref
          required: true
          schema: { type: string }
          description: |
            Reference returned by createChaos. Encodes namespace/name/kind.
      security:
        - internalToken: []
      responses:
        "204":
          description: deleted (or already gone)
        "401":
          description: bad internal token
  /api/targets:
    get:
      operationId: listTargets
      responses:
        "200":
          description: registered targets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Target" }
  /api/targets/{id}:
    get:
      operationId: getTarget
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: target detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Target" }
        "404":
          description: not found
  /api/targets/{id}/test:
    post:
      operationId: testTargetConnection
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: probe result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProbeResult" }
        "404":
          description: not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  schemas:
    Scenario:
      type: object
      required: [id, displayName]
      properties:
        id:           { type: string }
        displayName:  { type: string }
        description:  { type: string }
        targetType:   { type: string, description: "e.g. mysql, kafka, doris" }
        parameters:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:        { type: string }
              default:     { type: string }
              description: { type: string }
    Run:
      type: object
      required: [id, scenario, status, startedAt]
      properties:
        id:         { type: string }
        scenario:   { type: string }
        status:     { type: string, enum: [Pending, Running, Succeeded, Failed, Error, Unknown] }
        startedAt:  { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        score:      { type: number, format: double, nullable: true }
        workflowName: { type: string }
        target:       { type: string, description: "Remote target ID (Run was injected into a remote cluster). Empty = local." }
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          properties:
            parameters:
              type: object
              additionalProperties: { type: string }
            steps:
              type: array
              items:
                type: object
                required: [name, phase]
                properties:
                  name:      { type: string }
                  phase:     { type: string }
                  startedAt: { type: string, format: date-time }
                  finishedAt: { type: string, format: date-time }
                  message:   { type: string }
            verdict:
              type: object
              nullable: true
              description: "Decoded from MinIO report.json. Absent if no report yet."
              additionalProperties: true
            grafanaUrls:
              type: array
              items:
                type: object
                required: [label, url]
                properties:
                  label: { type: string }
                  url:   { type: string }
            target:
              type: string
              description: "Remote target ID (Run was injected into a remote cluster). Empty = local."
    CreateRunRequest:
      type: object
      required: [scenarioId]
      properties:
        scenarioId:
          type: string
          description: "WorkflowTemplate name (e.g. mysql-pod-delete)"
        targetId:
          type: string
          description: "Optional remote target ID. Empty = inject chaos in framework cluster."
        parameters:
          type: object
          description: "Optional parameter overrides. Keys are WT parameter names."
          additionalProperties:
            type: string
    ChaosResource:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion: { type: string, example: "chaos-mesh.org/v1alpha1" }
        kind:       { type: string, example: "Schedule" }
        metadata:
          type: object
          additionalProperties: true
        spec:
          type: object
          additionalProperties: true
    ChaosResourceRef:
      type: object
      required: [ref]
      properties:
        ref:        { type: string, description: "Opaque handle. URL-safe. Pass to DELETE /internal/chaos/{ref}." }
        kind:       { type: string }
        name:       { type: string }
        namespace:  { type: string }
    Target:
      type: object
      required: [id]
      properties:
        id:                 { type: string }
        displayName:        { type: string }
        kubeconfigSecret:   { type: string }
        allowedTargetTypes: { type: array, items: { type: string } }
        namespace:          { type: string }
        configured:         { type: boolean, description: "True if a valid kubeconfig was loaded for this target." }
    ProbeResult:
      type: object
      required: [ok]
      properties:
        ok:           { type: boolean }
        latencyNanos: { type: integer, format: int64 }
        error:        { type: string }
security:
  - bearerAuth: []