# controlplane/internal/api/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package gen

import (
	"time"
)

const (
	BearerAuthScopes    = "bearerAuth.Scopes"
	InternalTokenScopes = "internalToken.Scopes"
)

// Defines values for RunStatus.
const (
	RunStatusError     RunStatus = "Error"
	RunStatusFailed    RunStatus = "Failed"
	RunStatusPending   RunStatus = "Pending"
	RunStatusRunning   RunStatus = "Running"
	RunStatusSucceeded RunStatus = "Succeeded"
	RunStatusUnknown   RunStatus = "Unknown"
)

// Defines values for RunDetailStatus.
const (
	RunDetailStatusError     RunDetailStatus = "Error"
	RunDetailStatusFailed    RunDetailStatus = "Failed"
	RunDetailStatusPending   RunDetailStatus = "Pending"
	RunDetailStatusRunning   RunDetailStatus = "Running"
	RunDetailStatusSucceeded RunDetailStatus = "Succeeded"
	RunDetailStatusUnknown   RunDetailStatus = "Unknown"
)

// ChaosResource defines model for ChaosResource.
type ChaosResource struct {
	ApiVersion string                 `json:"apiVersion"`
	Kind       string                 `json:"kind"`
	Metadata   map[string]interface{} `json:"metadata"`
	Spec       map[string]interface{} `json:"spec"`
}

// ChaosResourceRef defines model for ChaosResourceRef.
type ChaosResourceRef struct {
	Kind      *string `json:"kind,omitempty"`
	Name      *string `json:"name,omitempty"`
	Namespace *string `json:"namespace,omitempty"`

	// Ref Opaque handle. URL-safe. Pass to DELETE /internal/chaos/{ref}.
	Ref string `json:"ref"`
}

// CreateRunRequest defines model for CreateRunRequest.
type CreateRunRequest struct {
	// Parameters Optional parameter overrides. Keys are WT parameter names.
	Parameters *map[string]string `json:"parameters,omitempty"`

	// ScenarioId WorkflowTemplate name (e.g. mysql-pod-delete)
	ScenarioId string `json:"scenarioId"`

	// TargetId Optional remote target ID. Empty = inject chaos in framework cluster.
	TargetId *string `json:"targetId,omitempty"`
}

// ProbeResult defines model for ProbeResult.
type ProbeResult struct {
	Error        *string `json:"error,omitempty"`
	LatencyNanos *int64  `json:"latencyNanos,omitempty"`
	Ok           bool    `json:"ok"`
}

// Run defines model for Run.
type Run struct {
	FinishedAt *time.Time `json:"finishedAt,omitempty"`
	Id         string     `json:"id"`
	Scenario   string     `json:"scenario"`
	Score      *float64   `json:"score"`
	StartedAt  time.Time  `json:"startedAt"`
	Status     RunStatus  `json:"status"`

	// Target Remote target ID (Run was injected into a remote cluster). Empty = local.
	Target       *string `json:"target,omitempty"`
	WorkflowName *string `json:"workflowName,omitempty"`
}

// RunStatus defines model for Run.Status.
type RunStatus string

// RunDetail defines model for RunDetail.
type RunDetail struct {
	FinishedAt  *time.Time `json:"finishedAt,omitempty"`
	GrafanaUrls *[]struct {
		Label string `json:"label"`
		Url   string `json:"url"`
	} `json:"grafanaUrls,omitempty"`
	Id         string             `json:"id"`
	Parameters *map[string]string `json:"parameters,omitempty"`
	Scenario   string             `json:"scenario"`
	Score      *float64           `json:"score"`
	StartedAt  time.Time          `json:"startedAt"`
	Status     RunDetailStatus    `json:"status"`
	Steps      *[]struct {
		FinishedAt *time.Time `json:"finishedAt,omitempty"`
		Message    *string    `json:"message,omitempty"`
		Name       string     `json:"name"`
		Phase      string     `json:"phase"`
		StartedAt  *time.Time `json:"startedAt,omitempty"`
	} `json:"steps,omitempty"`

	// Target Remote target ID (Run was injected into a remote cluster). Empty = local.
	Target *string `json:"target,omitempty"`

	// Verdict Decoded from MinIO report.json. Absent if no report yet.
	Verdict      *map[string]interface{} `json:"verdict"`
	WorkflowName *string                 `json:"workflowName,omitempty"`
}

// RunDetailStatus defines model for RunDetail.Status.
type RunDetailStatus string

// Scenario defines model for Scenario.
type Scenario struct {
	Description *string `json:"description,omitempty"`
	DisplayName string  `json:"displayName"`
	Id          string  `json:"id"`
	Parameters  *[]struct {
		Default     *string `json:"default,omitempty"`
		Description *string `json:"description,omitempty"`
		Name        string  `json:"name"`
	} `json:"parameters,omitempty"`

	// TargetType e.g. mysql, kafka, doris
	TargetType *string `json:"targetType,omitempty"`
}

// Target defines model for Target.
type Target struct {
	AllowedTargetTypes *[]string `json:"allowedTargetTypes,omitempty"`

	// Configured True if a valid kubeconfig was loaded for this target.
	Configured       *bool   `json:"configured,omitempty"`
	DisplayName      *string `json:"displayName,omitempty"`
	Id               string  `json:"id"`
	KubeconfigSecret *string `json:"kubeconfigSecret,omitempty"`
	Namespace        *string `json:"namespace,omitempty"`
}

// ListRunsParams defines parameters for ListRuns.
type ListRunsParams struct {
	Scenario *string    `form:"scenario,omitempty" json:"scenario,omitempty"`
	Status   *string    `form:"status,omitempty" json:"status,omitempty"`
	Since    *time.Time `form:"since,omitempty" json:"since,omitempty"`
	Limit    *int       `form:"limit,omitempty" json:"limit,omitempty"`
}

// CreateRunJSONRequestBody defines body for CreateRun for application/json ContentType.
type CreateRunJSONRequestBody = CreateRunRequest

// CreateChaosJSONRequestBody defines body for CreateChaos for application/json ContentType.
type CreateChaosJSONRequestBody = ChaosResource